@@ -6,20 +6,23 @@ import android.graphics.BitmapFactory
 import android.os.Bundle
 import android.util.Base64
 import android.view.View
+import android.widget.ImageView
+import android.widget.TextView
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import com.lkpc.android.app.glory.R
 import com.lkpc.android.app.glory.constants.SharedPreference
-import kotlinx.android.synthetic.main.action_bar.*
-import kotlinx.android.synthetic.main.activity_qr_code_generator.*
+import com.lkpc.android.app.glory.databinding.ActivityQrCodeGeneratorBinding
 import net.glxn.qrgen.android.QRCode
 import java.io.ByteArrayOutputStream
 
 
 class QrCodeGeneratorActivity : AppCompatActivity() {
+    private lateinit var binding: ActivityQrCodeGeneratorBinding
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_qr_code_generator)
+        binding = ActivityQrCodeGeneratorBinding.inflate(layoutInflater)
+        setContentView(binding.root)
         supportActionBar!!.setDisplayShowCustomEnabled(true)
         supportActionBar!!.setCustomView(R.layout.action_bar)
 
@@ -28,29 +31,29 @@ class QrCodeGeneratorActivity : AppCompatActivity() {
         val name = sp.getString(SharedPreference.QR_KEY_NAME, "")
         val phone = sp.getString(SharedPreference.QR_KEY_PHONE, "")
         val imgStr = sp.getString(SharedPreference.QR_KEY_IMAGE, "")
-        qr_name.setText(name)
-        qr_phone.setText(phone)
+        binding.qrName.setText(name)
+        binding.qrPhone.setText(phone)
         if (imgStr != null && imgStr.isNotEmpty()) {
             val imageAsBytes: ByteArray = Base64.decode(imgStr.encodeToByteArray(), Base64.DEFAULT)
-            img_qr_code.setImageBitmap(
+            binding.imgQrCode.setImageBitmap(
                 BitmapFactory.decodeByteArray(imageAsBytes, 0, imageAsBytes.size))
         }
 
-        ab_title.text = getString(R.string.qr_code_gen_title)
-        ab_btn_back.visibility = View.VISIBLE
-        ab_btn_back.setOnClickListener {
+        findViewById<TextView>(R.id.ab_title).text = getString(R.string.qr_code_gen_title)
+        findViewById<ImageView>(R.id.ab_btn_back).visibility = View.VISIBLE
+        findViewById<ImageView>(R.id.ab_btn_back).setOnClickListener {
             finish()
         }
 
-        btn_qr_code_generate.setOnClickListener {
-            if (qr_name.text.isEmpty() || qr_phone.text.isEmpty()) {
+        binding.btnQrCodeGenerate.setOnClickListener {
+            if (binding.qrName.text.isEmpty() || binding.qrPhone.text.isEmpty()) {
                 Toast.makeText(this, "이름 및 휴대폰 번호를 입력하세요", Toast.LENGTH_SHORT).show()
                 return@setOnClickListener
             }
 
-            val bm: Bitmap = QRCode.from("${qr_name.text} ${qr_phone.text}")
+            val bm: Bitmap = QRCode.from("${binding.qrName.text} ${binding.qrPhone.text}")
                 .withCharset("UTF-8").withSize(250, 250).bitmap()
-            img_qr_code.setImageBitmap(bm)
+            binding.imgQrCode.setImageBitmap(bm)
 
             // update sharedPreferences
             val outputStream = ByteArrayOutputStream()
@@ -61,8 +64,8 @@ class QrCodeGeneratorActivity : AppCompatActivity() {
                 Base64.encodeToString(b, Base64.DEFAULT)
             ).apply()
 
-            sp.edit().putString(SharedPreference.QR_KEY_NAME, "${qr_name.text}").apply()
-            sp.edit().putString(SharedPreference.QR_KEY_PHONE, "${qr_phone.text}").apply()
+            sp.edit().putString(SharedPreference.QR_KEY_NAME, "${binding.qrName.text}").apply()
+            sp.edit().putString(SharedPreference.QR_KEY_PHONE, "${binding.qrPhone.text}").apply()
         }
     }
 }
