@@ -5,39 +5,43 @@ import android.net.Uri
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.View
+import android.widget.ImageView
+import android.widget.TextView
 import com.lkpc.android.app.glory.R
 import com.lkpc.android.app.glory.constants.WebUrls
-import kotlinx.android.synthetic.main.action_bar.*
-import kotlinx.android.synthetic.main.activity_youtube_channel.*
+import com.lkpc.android.app.glory.databinding.ActivityYoutubeChannelBinding
 
 class YoutubeChannelActivity : AppCompatActivity() {
+    private lateinit var binding: ActivityYoutubeChannelBinding
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_youtube_channel)
-        supportActionBar!!.setDisplayShowCustomEnabled(true)
-        supportActionBar!!.setCustomView(R.layout.action_bar)
+        binding = ActivityYoutubeChannelBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+        supportActionBar?.setDisplayShowCustomEnabled(true)
+        supportActionBar?.setCustomView(R.layout.action_bar)
 
         // setup action bar
-        ab_title.text = getString(R.string.lpc_youtube_channels_kr)
-        ab_btn_back.visibility = View.VISIBLE
-        ab_btn_back.setOnClickListener {
+        findViewById<TextView>(R.id.ab_title).text = getString(R.string.lpc_youtube_channels_kr)
+        findViewById<ImageView>(R.id.ab_btn_back).visibility = View.VISIBLE
+        findViewById<ImageView>(R.id.ab_btn_back).setOnClickListener {
             finish()
         }
 
         // setup buttons
-        channel1.setOnClickListener {
+        binding.channel1.setOnClickListener {
             startActivity(Intent(Intent.ACTION_VIEW, Uri.parse(WebUrls.LKPC_CHANNEL)))
         }
-        channel2.setOnClickListener {
+        binding.channel2.setOnClickListener {
             startActivity(Intent(Intent.ACTION_VIEW, Uri.parse(WebUrls.YA_CHANNEL)))
         }
-        channel3.setOnClickListener {
+        binding.channel3.setOnClickListener {
             startActivity(Intent(Intent.ACTION_VIEW, Uri.parse(WebUrls.EC_CHANNEL)))
         }
-        channel4.setOnClickListener {
+        binding.channel4.setOnClickListener {
             startActivity(Intent(Intent.ACTION_VIEW, Uri.parse(WebUrls.NHF_CHANNEL)))
         }
-        channel5.setOnClickListener {
+        binding.channel5.setOnClickListener {
             startActivity(Intent(Intent.ACTION_VIEW, Uri.parse(WebUrls.DT_CHANNEL)))
         }
     }
