@@ -6,18 +6,25 @@ import android.content.Intent
 import android.net.Uri
 import android.os.Build
 import android.os.Bundle
+import android.os.PersistableBundle
+import android.view.MenuItem
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
 import androidx.core.view.GravityCompat
 import androidx.fragment.app.Fragment
+import androidx.navigation.NavController
+import androidx.navigation.Navigation
 import androidx.navigation.findNavController
+import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.ui.*
 import androidx.preference.PreferenceManager
+import com.google.android.material.navigation.NavigationView
 import com.google.firebase.messaging.FirebaseMessaging
 import com.lkpc.android.app.glory.constants.Notification
 import com.lkpc.android.app.glory.constants.Notification.Companion.CHANNEL_ID
 import com.lkpc.android.app.glory.constants.SharedPreference
 import com.lkpc.android.app.glory.constants.WebUrls
+import com.lkpc.android.app.glory.databinding.ActivityMainBinding
 import com.lkpc.android.app.glory.ui.basic_webview.BasicWebviewActivity
 import com.lkpc.android.app.glory.ui.bulletin.BulletinActivity
 import com.lkpc.android.app.glory.ui.calendar.CalendarActivity
@@ -33,10 +40,6 @@ import com.lkpc.android.app.glory.ui.note.NoteListActivity
 import com.lkpc.android.app.glory.ui.qr_code.QrCodeGeneratorActivity
 import com.lkpc.android.app.glory.ui.sermon.SermonFragment
 import com.lkpc.android.app.glory.ui.settings.SettingsActivity
-import kotlinx.android.synthetic.main.activity_main.*
-import kotlinx.android.synthetic.main.app_bar_main.*
-import kotlinx.android.synthetic.main.tool_bar.*
-
 
 class MainActivity : AppCompatActivity() {
     companion object {
@@ -97,15 +100,18 @@ class MainActivity : AppCompatActivity() {
     private var activeFragment  : Fragment? = null
 
     private lateinit var appBarConfiguration: AppBarConfiguration
+    private lateinit var binding: ActivityMainBinding
+    private lateinit var navController: NavController
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
         val toolbar: Toolbar = findViewById(R.id.toolbar)
         setSupportActionBar(toolbar)
         supportActionBar!!.setDisplayShowTitleEnabled(false)
 
-        toolbar_title.setText(R.string.lpc)
+        binding.appBarMain.toolBar.toolbarTitle.setText(R.string.lpc)
 
         // handle bundle data
         val data = intent.getStringExtra("contents")
@@ -130,8 +136,10 @@ class MainActivity : AppCompatActivity() {
         }
 
         if (savedInstanceState == null) {
-            supportFragmentManager.beginTransaction().detach(nav_host_fragment).commitNow()
-            //add all fragments but show only active fragment
+            supportFragmentManager.findFragmentById(R.id.nav_host_fragment)?.let {
+                supportFragmentManager.beginTransaction().detach(it).commitNow()
+            }
+            // add all fragments but show only active fragment
             supportFragmentManager.beginTransaction()
                 .add(R.id.nav_host_fragment, homeFragment, TAG_HOME).hide(homeFragment)
                 .add(R.id.nav_host_fragment, columnFragment, TAG_COLUMN).hide(columnFragment)
@@ -141,38 +149,36 @@ class MainActivity : AppCompatActivity() {
                 .show(activeFragment!!)
                 .commit()
         }
-
         setupNavigationView()
         setupNotification()
     }
 
     override fun onSupportNavigateUp(): Boolean {
-        val navController = findNavController(R.id.nav_host_fragment)
         return navController.navigateUp(appBarConfiguration) || super.onSupportNavigateUp()
     }
 
     override fun onBackPressed() {
-        if (drawer_layout.isOpen) {
-            drawer_layout.closeDrawer(GravityCompat.START)
+        if (binding.drawerLayout.isOpen) {
+            binding.drawerLayout.closeDrawer(GravityCompat.START)
         } else {
             super.onBackPressed()
         }
     }
 
     private fun setupNavigationView() {
-        val navController = findNavController(R.id.nav_host_fragment)
-
+        val navHostFragment = supportFragmentManager.findFragmentById(R.id.nav_host_fragment) as NavHostFragment
+        navController = navHostFragment.navController
         // setup drawer
         appBarConfiguration = AppBarConfiguration(setOf(
             R.id.navigation_home, R.id.navigation_column, R.id.navigation_sermon,
             R.id.navigation_meditation, R.id.navigation_news,
             R.id.nav_menu_qr_code, R.id.nav_menu_my_note, R.id.nav_menu_online_meet,
             R.id.nav_menu_church_events, R.id.nav_menu_service_info,
-            R.id.nav_menu_nav_guide), drawer_layout)
+            R.id.nav_menu_nav_guide), binding.drawerLayout)
         setupActionBarWithNavController(navController, appBarConfiguration)
-        drawer_nav_view.setupWithNavController(navController)
-        drawer_nav_view.setNavigationItemSelectedListener { item ->
-            drawer_layout.closeDrawer(GravityCompat.START)
+        binding.drawerNavView.setupWithNavController(navController)
+        binding.drawerNavView.setNavigationItemSelectedListener { item ->
+            binding.drawerLayout.closeDrawer(GravityCompat.START)
 
             when (item.itemId) {
                 R.id.nav_menu_qr_code -> {
@@ -237,12 +243,9 @@ class MainActivity : AppCompatActivity() {
         }
 
         // setup bottom navigation
-        bottom_nav_view.setupWithNavController(navController)
-        bottom_nav_view.setOnNavigationItemSelectedListener {
+        binding.appBarMain.bottomNavView.setupWithNavController(navController)
+        binding.appBarMain.bottomNavView.setOnItemSelectedListener {
             setFragment(it.itemId)
-        }
-        bottom_nav_view.setOnNavigationItemReselectedListener {
-            return@setOnNavigationItemReselectedListener
         }
     }
 
@@ -279,7 +282,7 @@ class MainActivity : AppCompatActivity() {
                     .show(homeFragment)
                     .commit()
                 activeFragment = homeFragment
-                toolbar_title.setText(R.string.lpc)
+                binding.appBarMain.toolBar.toolbarTitle.setText(R.string.lpc)
             }
             R.id.navigation_column  ->{
                 if (activeFragment is ColumnFragment) {
@@ -289,7 +292,7 @@ class MainActivity : AppCompatActivity() {
                     .show(columnFragment)
                     .commit()
                 activeFragment = columnFragment
-                toolbar_title.setText(R.string.title_column)
+                binding.appBarMain.toolBar.toolbarTitle.setText(R.string.title_column)
             }
             R.id.navigation_sermon ->{
                 if (activeFragment is SermonFragment) {
@@ -299,7 +302,7 @@ class MainActivity : AppCompatActivity() {
                     .show(sermonFragment)
                     .commit()
                 activeFragment = sermonFragment
-                toolbar_title.setText(R.string.title_sermon)
+                binding.appBarMain.toolBar.toolbarTitle.setText(R.string.title_sermon)
             }
             R.id.navigation_meditation ->{
                 if (activeFragment is MeditationFragment) {
@@ -309,7 +312,7 @@ class MainActivity : AppCompatActivity() {
                     .show(meditationFragment)
                     .commit()
                 activeFragment = meditationFragment
-                toolbar_title.setText(R.string.title_meditation)
+                binding.appBarMain.toolBar.toolbarTitle.setText(R.string.title_meditation)
             }
             R.id.navigation_news ->{
                 if (activeFragment is NewsFragment) {
@@ -318,7 +321,7 @@ class MainActivity : AppCompatActivity() {
                 supportFragmentManager.beginTransaction().hide(activeFragment!!)
                     .show(newsFragment).commit()
                 activeFragment = newsFragment
-                toolbar_title.setText(R.string.title_notifications)
+                binding.appBarMain.toolBar.toolbarTitle.setText(R.string.title_notifications)
             }
         }
         return true
