@@ -10,7 +10,6 @@ import com.lkpc.android.app.glory.R
 import com.lkpc.android.app.glory.constants.WebUrls
 import com.lkpc.android.app.glory.entity.BaseContent
 import com.lkpc.android.app.glory.ui.basic_webview.BasicWebviewActivity
-import kotlinx.android.synthetic.main.list_item_bulletin.view.*
 import java.text.SimpleDateFormat
 import java.util.*
 
@@ -23,8 +22,8 @@ class BulletinAdapter : RecyclerView.Adapter<RecyclerView.ViewHolder>() {
     var bulletins: List<BaseContent?> = mutableListOf()
 
     class ItemViewHolder(view: View): RecyclerView.ViewHolder(view) {
-        var title: TextView = view.bulletin_title
-        var date: TextView = view.bulletin_date
+        var title: TextView = view.findViewById(R.id.bulletin_title)
+        var date: TextView = view.findViewById(R.id.bulletin_date)
     }
 
     class LoadingViewHolder(view: View): RecyclerView.ViewHolder(view) {
