@@ -10,56 +10,58 @@ import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.Observer
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.lkpc.android.app.glory.R
+import com.lkpc.android.app.glory.databinding.FragmentColumnBinding
 import com.lkpc.android.app.glory.entity.BaseContent
-import kotlinx.android.synthetic.main.app_bar_main.*
-import kotlinx.android.synthetic.main.fragment_column.*
 
 class ColumnFragment : Fragment() {
+
+    private var _binding: FragmentColumnBinding? = null
+    // This property is only valid between onCreateView and onDestroyView.
+    private val binding get() = _binding!!
+
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.fragment_column, container, false)
+    ): View {
+        _binding = FragmentColumnBinding.inflate(inflater, container, false)
+        return binding.root
     }
 
-    override fun onActivityCreated(savedInstanceState: Bundle?) {
-        super.onActivityCreated(savedInstanceState)
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
 
-        rv_column.layoutManager = LinearLayoutManager(activity)
-        rv_column.adapter = ColumnAdapter()
+        binding.rvColumn.layoutManager = LinearLayoutManager(activity)
+        binding.rvColumn.adapter = ColumnAdapter()
 
         // data observation
         val viewModel: ColumnViewModel by viewModels()
         val observer = Observer<List<BaseContent?>> { data ->
-            if (rv_column != null) {
-                val adapter = rv_column.adapter as ColumnAdapter
-                if (adapter.isLoading) {
-                    (adapter.columns as MutableList<BaseContent?>).removeAt(adapter.columns.size - 1)
-                    adapter.isLoading = false
-                }
-                adapter.columns = data
-                adapter.notifyDataSetChanged()
+            val adapter = binding.rvColumn.adapter as ColumnAdapter
+            if (adapter.isLoading) {
+                (adapter.columns as MutableList<BaseContent?>).removeAt(adapter.columns.size - 1)
+                adapter.isLoading = false
             }
+            adapter.columns = data
+            adapter.notifyDataSetChanged()
         }
         viewModel.getData().observe(activity as LifecycleOwner, observer)
 
         // setup refresh
-        column_layout.setOnRefreshListener {
+        binding.columnLayout.setOnRefreshListener {
             viewModel.addData(0)
-            column_layout.isRefreshing = false
+            binding.columnLayout.isRefreshing = false
         }
 
         // scroll listener
-        rv_column.addOnScrollListener(object : RecyclerView.OnScrollListener() {
+        binding.rvColumn.addOnScrollListener(object : RecyclerView.OnScrollListener() {
             override fun onScrolled(recyclerView: RecyclerView, dx: Int, dy: Int) {
                 super.onScrolled(recyclerView, dx, dy)
 
-                val adapter = rv_column.adapter as ColumnAdapter
-                if (!rv_column.canScrollVertically(1) && !adapter.isLoading) {
+                val adapter = binding.rvColumn.adapter as ColumnAdapter
+                if (!binding.rvColumn.canScrollVertically(1) && !adapter.isLoading) {
                     (adapter.columns as MutableList).add(null)
                     adapter.notifyItemInserted(adapter.columns.size - 1)
-                    rv_column.scrollToPosition(adapter.columns.size - 1)
+                    binding.rvColumn.scrollToPosition(adapter.columns.size - 1)
 
                     viewModel.addData(adapter.itemCount - 1)
 
