@@ -2,39 +2,50 @@ package com.lkpc.android.app.glory.ui.fellow_news
 
 import android.os.Bundle
 import android.view.View
+import android.widget.ImageView
+import android.widget.TextView
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.Observer
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.lkpc.android.app.glory.R
+import com.lkpc.android.app.glory.databinding.ActivityFellowNewsBinding
 import com.lkpc.android.app.glory.entity.BaseContent
-import kotlinx.android.synthetic.main.action_bar.*
-import kotlinx.android.synthetic.main.activity_fellow_news.*
 
 class FellowNewsActivity : AppCompatActivity() {
+
+    private lateinit var binding: ActivityFellowNewsBinding
+
+    private var actionBarTitle: TextView? = null
+    private var actionBarBackBtn: ImageView? = null
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_fellow_news)
-        supportActionBar!!.setDisplayShowCustomEnabled(true)
-        supportActionBar!!.setCustomView(R.layout.action_bar)
+        binding = ActivityFellowNewsBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+        supportActionBar?.setDisplayShowCustomEnabled(true)
+        supportActionBar?.setCustomView(R.layout.action_bar)
 
-        ab_title.text = getString(R.string.fellow_news)
+        actionBarTitle = supportActionBar?.customView?.findViewById(R.id.ab_title)
+        actionBarBackBtn = supportActionBar?.customView?.findViewById(R.id.ab_btn_back)
 
-        ab_btn_back.visibility = View.VISIBLE
-        ab_btn_back.setOnClickListener {
+        actionBarTitle?.text = getString(R.string.fellow_news)
+
+        actionBarBackBtn?.visibility = View.VISIBLE
+        actionBarBackBtn?.setOnClickListener {
             finish()
         }
 
-        rv_fellow_news.layoutManager = LinearLayoutManager(this)
-        rv_fellow_news.adapter = FellowNewsAdapter()
+        binding.rvFellowNews.layoutManager = LinearLayoutManager(this)
+        binding.rvFellowNews.adapter = FellowNewsAdapter()
 
         // data observation
         val viewModel : FellowNewsViewModel by viewModels()
 
         val observer = Observer<List<BaseContent?>> { data ->
-            if (rv_fellow_news != null) {
-                val adapter = rv_fellow_news.adapter as FellowNewsAdapter
+            if (binding.rvFellowNews != null) {
+                val adapter = binding.rvFellowNews.adapter as FellowNewsAdapter
                 if (adapter.isLoading) {
                     (adapter.fellowNews as MutableList<BaseContent?>)
                         .removeAt(adapter.fellowNews.size - 1)
@@ -47,21 +58,21 @@ class FellowNewsActivity : AppCompatActivity() {
         viewModel.getData().observe(this, observer)
 
         // setup refresh
-        fellow_news_layout.setOnRefreshListener {
+        binding.fellowNewsLayout.setOnRefreshListener {
             viewModel.addData(0)
-            fellow_news_layout.isRefreshing = false
+            binding.fellowNewsLayout.isRefreshing = false
         }
 
         // scroll listener
-        rv_fellow_news.addOnScrollListener(object : RecyclerView.OnScrollListener() {
+        binding.rvFellowNews.addOnScrollListener(object : RecyclerView.OnScrollListener() {
             override fun onScrolled(recyclerView: RecyclerView, dx: Int, dy: Int) {
                 super.onScrolled(recyclerView, dx, dy)
 
-                val adapter = rv_fellow_news.adapter as FellowNewsAdapter
-                if (!rv_fellow_news.canScrollVertically(1) && !adapter.isLoading) {
+                val adapter = binding.rvFellowNews.adapter as FellowNewsAdapter
+                if (!binding.rvFellowNews.canScrollVertically(1) && !adapter.isLoading) {
                     (adapter.fellowNews as MutableList).add(null)
                     adapter.notifyItemInserted(adapter.fellowNews.size - 1)
-                    rv_fellow_news.scrollToPosition(adapter.fellowNews.size - 1)
+                    binding.rvFellowNews.scrollToPosition(adapter.fellowNews.size - 1)
 
                     viewModel.addData(adapter.itemCount - 1)
 
