@@ -10,13 +10,14 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import androidx.recyclerview.widget.RecyclerView
 import androidx.viewpager2.adapter.FragmentStateAdapter
 import com.lkpc.android.app.glory.R
 import com.lkpc.android.app.glory.api_client.BasicRequestClient
 import com.lkpc.android.app.glory.constants.WebUrls
+import com.lkpc.android.app.glory.databinding.HomeAdItemBinding
 import com.lkpc.android.app.glory.entity.AdContent
 import com.lkpc.android.app.glory.ui.basic_webview.BasicWebviewActivity
-import kotlinx.android.synthetic.main.home_ad_item.*
 import okhttp3.ResponseBody
 import retrofit2.Call
 import retrofit2.Callback
@@ -25,13 +26,22 @@ import retrofit2.Response
 class HomeSlidePagerAdapter(f: Fragment) : FragmentStateAdapter(f) {
     var adContents: MutableList<AdContent> = mutableListOf()
     var fragments: MutableList<Fragment> = mutableListOf()
+    private val ids: MutableList<Long> = mutableListOf()
 
     override fun getItemCount(): Int {
-        return adContents.count()
+        return adContents.size
     }
 
     override fun createFragment(position: Int): Fragment {
         return fragments[position]
+    }
+
+    override fun getItemId(position: Int): Long {
+        return ids[position]
+    }
+
+    override fun containsItem(itemId: Long): Boolean {
+        return ids.contains(itemId)
     }
 
     fun setData(ads : MutableList<AdContent>) {
@@ -44,21 +54,25 @@ class HomeSlidePagerAdapter(f: Fragment) : FragmentStateAdapter(f) {
                     override fun onResponse(call: Call<ResponseBody>, res: Response<ResponseBody>) {
                         if (res.isSuccessful && res.body() != null) {
                             f.img = BitmapFactory.decodeStream(res.body()!!.byteStream())
-                            f.refresh()
                         }
                     }
 
-                    override fun onFailure(call: Call<ResponseBody>?, t: Throwable?) {
-                        t?.printStackTrace()
+                    override fun onFailure(call: Call<ResponseBody>, t: Throwable) {
+                        t.printStackTrace()
                     }
                 })
             fragments.add(f)
+            ids.add(f.hashCode().toLong())
         }
         notifyDataSetChanged()
     }
 }
 
 class AdItemFragment(val content: AdContent) : Fragment() {
+    private var _binding: HomeAdItemBinding? = null
+    // This property is only valid between onCreateView and onDestroyView.
+    private val binding get() = _binding!!
+
     var img : Bitmap? = null
 
     override fun onCreateView(
@@ -66,18 +80,22 @@ class AdItemFragment(val content: AdContent) : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        return inflater.inflate(R.layout.home_ad_item, container, false)
+        _binding = HomeAdItemBinding.inflate(inflater, container, false)
+        return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        println("hansol onviewcreated")
         if (img != null) {
-            home_ad_item_img.setImageBitmap(img)
+            binding.homeAdItemImg.setImageBitmap(img)
 
             TransitionManager.beginDelayedTransition(
-                home_ad_item_layout as ViewGroup, Fade(Fade.IN))
-            home_ad_item_img.visibility = View.VISIBLE
+                binding.homeAdItemLayout as ViewGroup, Fade(Fade.IN))
+            binding.homeAdItemImg.visibility = View.VISIBLE
+
+            println("hansol onview 2")
         }
-        home_ad_item_layout.setOnClickListener {
+        binding.homeAdItemLayout.setOnClickListener {
             val i = Intent(view.context, BasicWebviewActivity::class.java)
             i.putExtra("title", R.string.banner_title)
             i.putExtra("url", WebUrls.LPC_HOMEPAGE + content.linkUrl)
@@ -86,12 +104,14 @@ class AdItemFragment(val content: AdContent) : Fragment() {
     }
 
     fun refresh() {
-        if (img != null && home_ad_item_img != null) {
-            home_ad_item_img.setImageBitmap(img)
+
+        println("hansol refresh")
+        if (img != null) {
+            binding.homeAdItemImg.setImageBitmap(img)
 
             TransitionManager.beginDelayedTransition(
-                home_ad_item_layout as ViewGroup, Fade(Fade.IN))
-            home_ad_item_img.visibility = View.VISIBLE
+                binding.homeAdItemLayout as ViewGroup, Fade(Fade.IN))
+            binding.homeAdItemImg.visibility = View.VISIBLE
         }
     }
 }
