@@ -2,14 +2,19 @@ package com.lkpc.android.app.glory.ui.location
 
 import android.os.Bundle
 import android.view.MenuItem
+import android.widget.ImageView
+import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
+import androidx.appcompat.widget.Toolbar
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.lkpc.android.app.glory.R
+import com.lkpc.android.app.glory.databinding.ActivityLocationBinding
 import com.lkpc.android.app.glory.entity.Location
-import kotlinx.android.synthetic.main.activity_location.*
-import kotlinx.android.synthetic.main.tool_bar.*
+import org.w3c.dom.Text
 
 class LocationActivity : AppCompatActivity() {
+    private lateinit var binding: ActivityLocationBinding
+
     private val locations = listOf(
         Location("미시사가", "6965 Professional Court, Mississauga, ON, L4V 1Y3",
         "Tel: 905-677-7729", "Fax: 905-677-7739", "Email: info@lkpc.org",
@@ -22,19 +27,23 @@ class LocationActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_location)
-        setSupportActionBar(toolbar)
+        binding = ActivityLocationBinding.inflate(layoutInflater)
+        setContentView(binding.root)
 
-        toolbar_title.setText(R.string.navigation)
+        val toolbar: Toolbar = findViewById(R.id.toolbar)
+        setSupportActionBar(toolbar)
 
         supportActionBar!!.setDisplayShowTitleEnabled(false)
         supportActionBar!!.setDisplayHomeAsUpEnabled(true)
 
+        val toolbarTitle: TextView? = supportActionBar?.customView?.findViewById(R.id.toolbar_title)
+        toolbarTitle?.setText(R.string.navigation)
+
         val adapter = LocationAdapter()
         adapter.setLocations(locations)
 
-        rv_location.layoutManager = LinearLayoutManager(this)
-        rv_location.adapter = adapter
+        binding.rvLocation.layoutManager = LinearLayoutManager(this)
+        binding.rvLocation.adapter = adapter
     }
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
