@@ -10,59 +10,59 @@ import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.Observer
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.lkpc.android.app.glory.R
+import com.lkpc.android.app.glory.databinding.FragmentNewsBinding
 import com.lkpc.android.app.glory.entity.BaseContent
-import kotlinx.android.synthetic.main.app_bar_main.*
-import kotlinx.android.synthetic.main.fragment_news.*
 
 
 class NewsFragment : Fragment() {
+
+    private var _binding: FragmentNewsBinding? = null
+    // This property is only valid between onCreateView and onDestroyView.
+    private val binding get() = _binding!!
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_news, container, false)
+        _binding = FragmentNewsBinding.inflate(inflater, container, false)
+        return binding.root
     }
 
-    override fun onActivityCreated(savedInstanceState: Bundle?) {
-        super.onActivityCreated(savedInstanceState)
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
 
-        rv_news.layoutManager = LinearLayoutManager(activity)
-        rv_news.adapter = NewsAdapter()
+        binding.rvNews.layoutManager = LinearLayoutManager(activity)
+        binding.rvNews.adapter = NewsAdapter()
 
         // data observation
         val viewModel: NewsViewModel by viewModels()
         val observer = Observer<List<BaseContent?>> { data ->
-            if (rv_news != null) {
-                val adapter = rv_news.adapter as NewsAdapter
-                if (adapter.isLoading) {
-                    (adapter.newsList as MutableList<BaseContent?>).removeAt(adapter.newsList.size - 1)
-                    adapter.isLoading = false
-                }
-                adapter.newsList = data
-                adapter.notifyDataSetChanged()
+            val adapter = binding.rvNews.adapter as NewsAdapter
+            if (adapter.isLoading) {
+                (adapter.newsList as MutableList<BaseContent?>).removeAt(adapter.newsList.size - 1)
+                adapter.isLoading = false
             }
+            adapter.newsList = data
+            adapter.notifyDataSetChanged()
         }
         viewModel.getData().observe(activity as LifecycleOwner, observer)
 
         // setup refresh
-        news_layout.setOnRefreshListener {
+        binding.newsLayout.setOnRefreshListener {
             viewModel.addData(0)
-            news_layout.isRefreshing = false
+            binding.newsLayout.isRefreshing = false
         }
 
         // scroll listener
-        rv_news.addOnScrollListener(object : RecyclerView.OnScrollListener() {
+        binding.rvNews.addOnScrollListener(object : RecyclerView.OnScrollListener() {
             override fun onScrolled(recyclerView: RecyclerView, dx: Int, dy: Int) {
                 super.onScrolled(recyclerView, dx, dy)
 
-                val adapter = rv_news.adapter as NewsAdapter
-                if (!rv_news.canScrollVertically(1) && !adapter.isLoading) {
+                val adapter = binding.rvNews.adapter as NewsAdapter
+                if (!binding.rvNews.canScrollVertically(1) && !adapter.isLoading) {
                     (adapter.newsList as MutableList).add(null)
                     adapter.notifyItemInserted(adapter.newsList.size - 1)
-                    rv_news.scrollToPosition(adapter.newsList.size - 1)
+                    binding.rvNews.scrollToPosition(adapter.newsList.size - 1)
 
                     viewModel.addData(adapter.itemCount - 1)
 
