@@ -4,25 +4,28 @@ import android.content.Intent
 import android.os.Bundle
 import android.view.Menu
 import android.view.MenuItem
+import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
 import com.lkpc.android.app.glory.R
 import com.lkpc.android.app.glory.data.NoteDatabase
+import com.lkpc.android.app.glory.databinding.ActivityNoteDetailBinding
 import com.lkpc.android.app.glory.entity.Note
-import kotlinx.android.synthetic.main.activity_note_detail.*
-import kotlinx.android.synthetic.main.tool_bar.*
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
 import java.text.SimpleDateFormat
 import java.util.*
 
 class NoteDetailActivity : AppCompatActivity() {
+    private lateinit var binding: ActivityNoteDetailBinding
+
     private var noteId : Int = -1
     private lateinit var note : Note
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_note_detail)
+        binding = ActivityNoteDetailBinding.inflate(layoutInflater)
+        setContentView(binding.root)
         val toolbar: Toolbar = findViewById(R.id.toolbar)
         setSupportActionBar(toolbar)
 
@@ -32,24 +35,24 @@ class NoteDetailActivity : AppCompatActivity() {
         noteId = intent.getIntExtra("noteId", -1)
         val db = NoteDatabase.getDatabase(context = this)
         db.noteDao().loadById(noteId).observe(
-            this, { note ->
-                if (note == null) {
-                    return@observe
-                }
-
-                this.note = note
-
-                // title
-                toolbar_title.text = note.title
-
-                // main content
-                note_detail_title.text = note.title
-                val dateFormat = SimpleDateFormat("yyyy-MM-dd HH:mm", Locale.CANADA)
-                val lastModified = dateFormat.format(note.lastModified)
-                note_detail_last_modified.text = lastModified
-                note_detail_content.text = note.content
+            this
+        ) { note ->
+            if (note == null) {
+                return@observe
             }
-        )
+
+            this.note = note
+
+            // title
+            toolbar.findViewById<TextView>(R.id.toolbar_title).text = note.title
+
+            // main content
+            binding.noteDetailTitle.text = note.title
+            val dateFormat = SimpleDateFormat("yyyy-MM-dd HH:mm", Locale.CANADA)
+            val lastModified = dateFormat.format(note.lastModified)
+            binding.noteDetailLastModified.text = lastModified
+            binding.noteDetailContent.text = note.content
+        }
     }
 
     override fun onCreateOptionsMenu(menu: Menu?): Boolean {
